refactor: Performance test code
--story=1
