+++ shardingdb_test.go
@@ -20,7 +20,6 @@ import (
 	"fmt"
 	"os"
 	"path"
-	"sync"
 	"testing"
 	"time"
 
@@ -285,166 +284,4 @@ func TestShardingDb_Iterator(t *testing.T) {
 	}
 	assert.Equal(t, 20, count)
 
-}
-func BenchmarkShardingDb_Put(b *testing.B) {
-	db := initDb(3)
-	defer db.Close()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		db.Put([]byte(fmt.Sprintf("key-%03d", i)), []byte(fmt.Sprintf("value-%03d", i)), nil)
-	}
-}
-func BenchmarkLevledb_Put(b *testing.B) {
-	db, err := leveldb.OpenFile(getTempDir(), nil)
-	assert.NoError(b, err)
-	defer db.Close()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		db.Put([]byte(fmt.Sprintf("key-%03d", i)), []byte(fmt.Sprintf("value-%03d", i)), nil)
-	}
-}
-func BenchmarkShardingDb_Get(b *testing.B) {
-	db := initDb(3)
-	defer db.Close()
-	for i := 0; i < 100000; i++ {
-		db.Put([]byte(fmt.Sprintf("key-%03d", i)), []byte(fmt.Sprintf("value-%03d", i)), nil)
-	}
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		db.Get([]byte(fmt.Sprintf("key-%03d", i)), nil)
-	}
-}
-func BenchmarkLeveldb_Get(b *testing.B) {
-	db, err := leveldb.OpenFile(getTempDir(), nil)
-	assert.NoError(b, err)
-	defer db.Close()
-	for i := 0; i < 100000; i++ {
-		db.Put([]byte(fmt.Sprintf("key-%03d", i)), []byte(fmt.Sprintf("value-%03d", i)), nil)
-	}
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		db.Get([]byte(fmt.Sprintf("key-%03d", i)), nil)
-	}
-}
-func TestShardingDbPerformance(t *testing.T) {
-	os.RemoveAll("/data/leveldb1")
-	os.RemoveAll("/data1/leveldb2")
-	dir3 := getTempDir()
-	thread := 100
-	loop := 100
-	batchSize := 1000
-	db, _ := OpenFile([]string{"/data/leveldb1", "/data1/leveldb2", dir3}, nil)
-	defer db.Close()
-	wg := sync.WaitGroup{}
-	wg.Add(thread)
-	start := time.Now()
-	for i := 0; i < thread; i++ {
-		go func(thr int) {
-			defer wg.Done()
-			for j := 0; j < loop; j++ {
-				batch := new(leveldb.Batch)
-				for k := 0; k < batchSize; k++ {
-					batch.Put([]byte(fmt.Sprintf("key-%02d-%03d", thr, k)), []byte(fmt.Sprintf("value-%03d", k)))
-				}
-				err := db.Write(batch, nil)
-				assert.NoError(t, err)
-			}
-		}(i)
-	}
-	wg.Wait()
-	fmt.Printf("ShardingDb write batch[%d] thread[%d] loop[%d] cost:%v\n", batchSize, thread, loop, time.Now().Sub(start))
-	//Test Get performance
-	wg = sync.WaitGroup{}
-	wg.Add(thread)
-	start = time.Now()
-	for i := 0; i < thread; i++ {
-		go func(thr int) {
-			defer wg.Done()
-			for j := 0; j < loop; j++ {
-				for k := 0; k < batchSize; k++ {
-					_, err := db.Get([]byte(fmt.Sprintf("key-%02d-%03d", thr, k)), nil)
-					assert.NoError(t, err)
-				}
-			}
-		}(i)
-	}
-	wg.Wait()
-	fmt.Printf("ShardingDb get batch[%d] thread[%d] loop[%d] cost:%v\n", batchSize, thread, loop, time.Now().Sub(start))
-	//Test Get not found performance
-	wg = sync.WaitGroup{}
-	wg.Add(thread)
-	start = time.Now()
-	for i := 0; i < thread; i++ {
-		go func(thr int) {
-			defer wg.Done()
-			for j := 0; j < loop; j++ {
-				for k := 0; k < batchSize; k++ {
-					v, _ := db.Get([]byte(fmt.Sprintf("key-%02d-%03d", thr, k+batchSize)), nil)
-					assert.Nil(t, v)
-				}
-			}
-		}(i)
-	}
-	wg.Wait()
-	fmt.Printf("ShardingDb get not found batch[%d] thread[%d] loop[%d] cost:%v\n", batchSize, thread, loop, time.Now().Sub(start))
-}
-func TestLeveldbPerformance(t *testing.T) {
-	thread := 100
-	loop := 100
-	batchSize := 1000
-	os.RemoveAll("/data/leveldb")
-	db, _ := leveldb.OpenFile("/data/leveldb", nil)
-	defer db.Close()
-	wg := sync.WaitGroup{}
-	wg.Add(thread)
-	start := time.Now()
-	for i := 0; i < thread; i++ {
-		go func(thr int) {
-			defer wg.Done()
-			for j := 0; j < loop; j++ {
-				batch := new(leveldb.Batch)
-				for k := 0; k < batchSize; k++ {
-					batch.Put([]byte(fmt.Sprintf("key-%02d-%03d", thr, k)), []byte(fmt.Sprintf("value-%03d", k)))
-				}
-				err := db.Write(batch, nil)
-				assert.NoError(t, err)
-			}
-		}(i)
-	}
-	wg.Wait()
-	fmt.Printf("Leveldb write batch[%d] thread[%d] loop[%d] cost:%v\n", batchSize, thread, loop, time.Now().Sub(start))
-	//Test Get performance
-	wg = sync.WaitGroup{}
-	wg.Add(thread)
-	start = time.Now()
-	for i := 0; i < thread; i++ {
-		go func(thr int) {
-			defer wg.Done()
-			for j := 0; j < loop; j++ {
-				for k := 0; k < batchSize; k++ {
-					_, err := db.Get([]byte(fmt.Sprintf("key-%02d-%03d", thr, k)), nil)
-					assert.NoError(t, err)
-				}
-			}
-		}(i)
-	}
-	wg.Wait()
-	fmt.Printf("Leveldb get batch[%d] thread[%d] loop[%d] cost:%v\n", batchSize, thread, loop, time.Now().Sub(start))
-	//Test Get not found performance
-	wg = sync.WaitGroup{}
-	wg.Add(thread)
-	start = time.Now()
-	for i := 0; i < thread; i++ {
-		go func(thr int) {
-			defer wg.Done()
-			for j := 0; j < loop; j++ {
-				for k := 0; k < batchSize; k++ {
-					v, _ := db.Get([]byte(fmt.Sprintf("key-%02d-%03d", thr, k+batchSize)), nil)
-					assert.Nil(t, v)
-				}
-			}
-		}(i)
-	}
-	wg.Wait()
-	fmt.Printf("Leveldb get not found batch[%d] thread[%d] loop[%d] cost:%v\n", batchSize, thread, loop, time.Now().Sub(start))
 }
